Implemented ordering while searching

+++ src/main/java/com/animalmanagementsystem/shelter/services/impl/AnimalServiceImpl.java
@@ -26,6 +26,7 @@ public class AnimalServiceImpl implements AnimalService {
     private final AnimalRepository animalRepository;
     private final AnimalMapper animalMapper;
     private final EntityManager entityManager;
+    public static final String NAME = "name";
 
     public AnimalServiceImpl(AnimalRepository animalRepository, AnimalMapper animalMapper, EntityManager entityManager) {
         this.animalRepository = animalRepository;
@@ -42,13 +43,16 @@ public class AnimalServiceImpl implements AnimalService {
 
         if (request.query() != null && !request.query().isBlank()) {
             String query = "%" + request.query() + "%";
-            Predicate namePredicate = criteriaBuilder.like(root.get("name"), query);
+            Predicate namePredicate = criteriaBuilder.like(root.get(NAME), query);
             Predicate speciesPredicate = criteriaBuilder.like(root.get("species"), query);
 
             predicates.add(criteriaBuilder.or(namePredicate, speciesPredicate));
         }
 
         criteriaQuery.where(criteriaBuilder.or(predicates.toArray(new Predicate[0])));
+
+        criteriaQuery.orderBy(criteriaBuilder.asc(root.get(NAME)));
+
         TypedQuery<AnimalEntity> query = entityManager.createQuery(criteriaQuery);
 
         return query.getResultList().stream().map(animalMapper::mapEntityToDto).toList();
@@ -95,7 +99,9 @@ public class AnimalServiceImpl implements AnimalService {
     public void deleteAnimal(Long id) {
         Optional<AnimalEntity> optionalAnimalEntity = animalRepository.findById(id);
 
-        optionalAnimalEntity.orElseThrow(() -> new AnimalNotFoundException(id));
+        if (optionalAnimalEntity.isEmpty()) {
+            throw new AnimalNotFoundException(id);
+        }
 
         animalRepository.deleteById(id);
     }
