+++ src/main/java/com/animalmanagementsystem/shelter/services/impl/CageServiceImpl.java
@@ -26,6 +26,7 @@ public class CageServiceImpl implements CageService {
     private final CageRepository cageRepository;
     private final CageMapper cageMapper;
     private final EntityManager entityManager;
+    public static final String CAGE_NUMBER = "cageNumber";
 
     public CageServiceImpl(CageRepository cageRepository, CageMapper cageMapper, EntityManager entityManager) {
         this.cageRepository = cageRepository;
@@ -42,13 +43,15 @@ public class CageServiceImpl implements CageService {
 
         if (request.query() != null && !request.query().isBlank()) {
             String query = "%" + request.query() + "%";
-            Predicate cageNumberPredicate = criteriaBuilder.like(root.get("cageNumber"), query);
+            Predicate cageNumberPredicate = criteriaBuilder.like(root.get(CAGE_NUMBER), query);
             Predicate availabilityPredicate = criteriaBuilder.like(root.get("availability"), query);
 
             predicates.add(criteriaBuilder.or(cageNumberPredicate, availabilityPredicate));
         }
 
         criteriaQuery.where(criteriaBuilder.or(predicates.toArray(new Predicate[0])));
+
+        criteriaQuery.orderBy(criteriaBuilder.asc(root.get(CAGE_NUMBER)));
 
         TypedQuery<CageEntity> query = entityManager.createQuery(criteriaQuery);
 
@@ -93,7 +96,9 @@ public class CageServiceImpl implements CageService {
     public void deleteCage(Long id) {
         Optional<CageEntity> optionalCageEntity = cageRepository.findById(id);
 
-        optionalCageEntity.orElseThrow(() -> new CageNotFoundException(id));
+        if (optionalCageEntity.isEmpty()) {
+            throw new CageNotFoundException(id);
+        }
 
         cageRepository.deleteById(id);
     }
